Logging fixes, addded channel support

+++ install.sh
@@ -1656,10 +1656,12 @@ watchdog_check() {
 	fi
 	# fi
 
-	FAILSIZE=$(wc -l <$WATCHDOG_FAILURES_FILE)
-	if [ "$FAILSIZE" -gt 10000 ]; then
-		tail -10000 $WATCHDOG_FAILURES_FILE >"$WATCHDOG_FAILURES_FILE.tmp"
-		mv -f "$WATCHDOG_FAILURES_FILE.tmp" "$WATCHDOG_FAILURES_FILE"
+	if [[ -f $WATCHDOG_FAILURES_FILE ]]; then
+		FAILSIZE=$(wc -l <$WATCHDOG_FAILURES_FILE)
+		if [ "$FAILSIZE" -gt 10000 ]; then
+			tail -10000 $WATCHDOG_FAILURES_FILE >"$WATCHDOG_FAILURES_FILE.tmp"
+			mv -f "$WATCHDOG_FAILURES_FILE.tmp" "$WATCHDOG_FAILURES_FILE"
+		fi
 	fi
 
 	echo "$WATCHDOG_TIME" >$WATCHDOG_TIME_FILE
@@ -1685,7 +1687,7 @@ parse_commandline_args() {
 
 	load_update_channel
 
-	CMDARGS=$(getopt --quiet -o wvdp --long watchdog,version,develop,production -n 'nightscout-tool' -- "$@")
+	CMDARGS=$(getopt --quiet -o wvdpc: --long watchdog,version,develop,production,channel: -n 'nightscout-tool' -- "$@")
 
 	# shellcheck disable=SC2181
 	if [ $? != 0 ]; then
@@ -1716,6 +1718,20 @@ parse_commandline_args() {
 		-p | --production)
 			warn "Switching to PRODUCTION update channel"
 			UPDATE_CHANNEL=master
+			echo "$UPDATE_CHANNEL" >$UPDATE_CHANNEL_FILE
+			shift
+			;;
+		-c | --channel)
+			shift # The arg is next in position args
+			UPDATE_CHANNEL_CANDIDATE=$1
+
+			[[ ! "$UPDATE_CHANNEL_CANDIDATE" =~ ^[a-z]{3,}$ ]] && {
+				echo "Incorrect channel name provided: $UPDATE_CHANNEL_CANDIDATE"
+				exit 1
+			}
+
+			warn "Switching to $UPDATE_CHANNEL_CANDIDATE update channel"
+			UPDATE_CHANNEL="$UPDATE_CHANNEL_CANDIDATE"
 			echo "$UPDATE_CHANNEL" >$UPDATE_CHANNEL_FILE
 			shift
 			;;
